quantum_aegis_backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
httpx==0.27.2
redis==5.0.7
rq==1.16.2
apscheduler==3.10.4
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
cryptography==43.0.1
